Made relation and from context user is validated

+++ todos/router.py
@@ -6,6 +6,8 @@ from common.responses.responses import success_response
 from common.exceptions.httpexception import http_exception
 from .schemas.todo import Todo
 from .schemas.todo_update import UpdateTodo
+from auth.context.current_user import get_current_user
+from auth.exceptions.user_exception import get_user_exception
 
 
 router = APIRouter(
@@ -21,16 +23,22 @@ async def read_all(db: Session = Depends(get_db)):
 
 
 @router.get("/{todo_id}")
-async def read_todo(todo_id: int, db: Session = Depends(get_db)):
-    todo = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(todo_id: int, user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
+    if user is None:
+        raise get_user_exception()
+    todo = db.query(Todos).filter(Todos.id == todo_id).filter(
+        Todos.owner_id == user.get('id')).first()
     if todo is None:
         raise http_exception(404,  f'Todo with {todo_id} not found')
     return todo
 
 
 @router.post("/")
-async def create_todo(todo: Todo, db: Session = Depends(get_db)):
+async def create_todo(todo: Todo, user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
+    if user is None:
+        raise get_user_exception()
     todo_model = Todos(**todo.model_dump())
+    todo_model.owner_id = user.get("id")
     db.add(todo_model)
     db.commit()
 
@@ -38,8 +46,11 @@ async def create_todo(todo: Todo, db: Session = Depends(get_db)):
 
 
 @router.put('/{todo_id}')
-async def update_todo(todo_id: int,  todo: UpdateTodo, db: Session = Depends(get_db)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id)
+async def update_todo(todo_id: int,  todo: UpdateTodo, user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
+    if user is None:
+        raise get_user_exception()
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(
+        Todos.owner_id == user.get('id'))
     todo_model_found = todo_model.first()
     if todo_model_found is None:
         raise http_exception(404, f'Todo with id {todo_id} not found')
@@ -52,8 +63,11 @@ async def update_todo(todo_id: int,  todo: UpdateTodo, db: Session = Depends(get
 
 
 @router.delete("/{todo_id}")
-async def delete_todo(todo_id: int, db: Session = Depends(get_db)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(todo_id: int, user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
+    if user is None:
+        raise get_user_exception()
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(
+        Todos.owner_id == user.get("id")).first()
     if todo_model is None:
         raise http_exception(404,  f'Todo with id {todo_id} not found')
     db.query(Todos).filter(Todos.id == todo_id).delete()
